[androsdk-1541] Rename package name
