+++ core/src/androidTest/java/org/hisp/dhis/android/realservertests/apischema/ApiSchema.kt
@@ -26,18 +26,16 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.realservertests.schema
+package org.hisp.dhis.android.realservertests.apischema
 
-import io.reactivex.Single
-import org.hisp.dhis.android.core.arch.api.fields.internal.Fields
-import org.hisp.dhis.android.core.arch.api.filters.internal.Which
-import org.hisp.dhis.android.core.arch.api.payload.internal.Payload
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-internal interface ApiSchemaService {
-    @GET("schemas")
-    fun getSchema(
-        @Query("fields") @Which fields: Fields<ApiSchema>
-    ): Single<Payload<ApiSchema>>
+internal data class ApiSchema(
+    val properties: List<SchemaProperty>
+) {
+    internal companion object {
+        internal data class SchemaProperty(
+            val propertyType: String,
+            val klass: String,
+            val constants: List<String>?
+        )
+    }
 }
