+++ core/src/androidTest/java/org/hisp/dhis/android/realservertests/apischema/ApiSchemaCall.kt
@@ -25,7 +25,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.realservertests.schema
+package org.hisp.dhis.android.realservertests.apischema
 
 import io.reactivex.Single
 import org.hisp.dhis.android.core.arch.call.factories.internal.ListCall
